% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{annotate_and_filter}
\alias{annotate_and_filter}
\title{Add in more useful columns, annotation and filter sites.}
\usage{
annotate_and_filter(dat, het_limit = 0.25, min_other_ratio = 0.15,
  min_depth = 5, max_sd_depth = 3)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP}

\item{het_limit}{limit fraction to call a heterozygote site.}

\item{min_other_ratio}{If the ratio of highest depth non-SNP allele is at
as high as this number, the SNP will be filtered. These SNPs either have
unusually high error rate (or mutation), or has a hidden (multi-allelic)
SNP. In either case, this position should be filtered from contamination
analysis.}

\item{min_depth}{if a SNP has depth less than this value, it will be
filtered. This metric is mainly for WGS where overall coverage is low.}

\item{max_sd_depth}{maximum number of standard deviations the depth
of a given SNP is less than the mean depth across positions. If a
position has depth less than that, it will be filtered. This metric is
mainly for positions with high depth such as targeted or exome sequencing
experiments.}
}
\value{
data.frame containing counts and metrics per SNP
}
\description{
Add in more useful columns, annotation and filter sites.
}
