% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{bayesian_genotype}
\alias{bayesian_genotype}
\title{Add bayesian_gt estimate of genotype based upon per-snp error rate and major and
minor allele counts/frequencies.}
\usage{
bayesian_genotype(dat)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP}
}
\value{
data.table containing counts and metrics per SNP
}
\description{
Add bayesian_gt estimate of genotype based upon per-snp error rate and major and
minor allele counts/frequencies.
}
