% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{bounded_freqs}
\alias{bounded_freqs}
\title{Bounds allele frequencies based on a specified amount}
\usage{
bounded_freqs(afs, pseudocount_freq)
}
\arguments{
\item{afs}{allele frequency}

\item{pseudocount_freq}{value to bound allele frequencies by}
}
\value{
bounded allele frequencies
}
\description{
Adds in allele count frequency or removes an allele count frequency
based on specificied value.
}
