% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{calculate_concordance}
\alias{calculate_concordance}
\title{Calculate concordance for all datasets}
\usage{
calculate_concordance(mt, i, d, cutoff, signature_SNPs)
}
\arguments{
\item{mt}{swap table to write to}

\item{i}{location on swap table}

\item{d}{list of data tables}

\item{cutoff}{cutoff for concordance}

\item{signature_SNPs}{table containing rsid, ref and alt of signature SNPs}
}
\description{
Calculate concordance for all datasets
}
