% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_error_model.R
\name{calculate_error_model}
\alias{calculate_error_model}
\title{Calculate error model}
\usage{
calculate_error_model(dat, save.dir = NA, sample = NA,
  default_error_rate = 1e-06)
}
\arguments{
\item{dat}{data.frame containing counts and metrics per SNP and only hom
alleles (0/0 and 1/1). Het alleles should have already been removed.}

\item{save.dir}{character location to save the results}

\item{sample}{character sample name}

\item{default_error_rate}{if absolutely no errors are observed in the data,
use this number as the error rate}
}
\value{
data.frame with error model
}
\description{
Calculate error model for ref/ref and alt/alt alleles globally across
all positions that should have no trace of contamination.
}
