% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{calculate_ranked_predictions}
\alias{calculate_ranked_predictions}
\title{Predict ethnicity per chromosome}
\usage{
calculate_ranked_predictions(initial_snp_likelihoods)
}
\arguments{
\item{initial_snp_likelihoods}{Data frame of the initial likelihoods per
snp per 1000G populations by
'get_snp_likelihoods()'}
}
\value{
Data frame with the log likelihoods per chromosome
}
\description{
This function produces ranked ethnicity predictions per chromosome. First,
takes in population likelihoods per snp as output from 'get_snp_likelihoods()'.
Second, calculates log likelihoods, and then log likelihood per chromosome.
Third, applies ethnicity priors to the likelihood values to generate the
posterior likelihoods. Lastly, calculates the P(data) and apply it to the posterior
likelihoods to generate the posterior probabilities per chromosome.
}
