% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{chr_stats}
\alias{chr_stats}
\title{Get the stats for a given chromosome from long biometrics file, including:
 - Mapq60 read count
 - Normalized mapq60 read count (normalized by total reads for this sample)
 - Fraction of bases covered by at least 1 read}
\usage{
chr_stats(biometrics_file, chr_name)
}
\arguments{
\item{biometrics_file}{input tsv file name}

\item{chr_name}{name of the chromosome to be counted}
}
\value{
list of chromosome metrics
}
\description{
Get the stats for a given chromosome from long biometrics file, including:
 - Mapq60 read count
 - Normalized mapq60 read count (normalized by total reads for this sample)
 - Fraction of bases covered by at least 1 read
}
