% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_wgs_functions.R
\name{clean}
\alias{clean}
\title{Clean given object and perform garbage collection}
\usage{
clean(object)
}
\arguments{
\item{object}{to be cleaned}
}
\description{
Clean given object and perform garbage collection
}
