% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{combine_counts}
\alias{combine_counts}
\title{Combine two count data tables}
\usage{
combine_counts(dat, dat2)
}
\arguments{
\item{dat}{data.table counts from first file}

\item{dat2}{data.table counts from second file}
}
\value{
data.table containing merged counts
}
\description{
Combine two count data tables
}
