% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{conta_ethnicity}
\alias{conta_ethnicity}
\title{Run the conta ethnicity prediction workflow}
\usage{
conta_ethnicity(gt_loh_file_loc, vcf, outdir)
}
\arguments{
\item{gt_loh_file_loc}{sample gt.loh.tsv file}

\item{vcf}{1000G vcf}

\item{outdir}{Path to directory to write outputs to.}
}
\value{
None, writes out TSVs and plots to `outdir`.
}
\description{
Runs the full conta ethnicity prediction workflow using a sample genotype
file with loh regions removed and 1000G vcf file. Workflow consists of
reading in, parsing, and preparing 1000G vcf file, followed by taking
the snp intersect with a samples gt.loh.tsv file. Ethnicity probabilities are
calculated per snp per chromosome. Generates tables and plots to the specified
output directory.
}
