% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{conta_lr}
\alias{conta_lr}
\title{Calculate a set of likelihood metrics across SNPs}
\usage{
conta_lr(cf, dat, EE, save.dir = NA, sample = NA)
}
\arguments{
\item{cf}{numeric contamination fraction to be tested}

\item{dat}{data.table containing counts and metrics per SNP, hets filtered}

\item{EE}{data.frame error model}

\item{save.dir}{directory to save results}

\item{sample}{name of the sample for printing purposes}
}
\value{
numeric avg. log-likelihood ratio for the given cf or a more detailed
  result if the save.dir and sample name are specified.
}
\description{
Calculate a set of likelihood metrics across SNPs
}
