% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_main.R
\name{conta_main}
\alias{conta_main}
\title{Run conta.}
\usage{
conta_main(tsv_file, sample, save_dir, metrics_file = "", lr_th = 0.04,
  sim_level = 0, baseline = NA, min_depth = 5, max_depth = 10000,
  loh_cutoff = 0.01, loh_delta_cutoff = 0.05, min_maf = 0.25,
  cf_correction = 0, min_cf = 0.00025, blackswan = 0.05,
  outlier_frac = 0.01, cores = 2)
}
\arguments{
\item{tsv_file}{input tsv file name}

\item{sample}{name of experiment (basename of file usually)}

\item{save_dir}{folder where the output will go to}

\item{metrics_file}{input tsv metrics file in long format}

\item{lr_th}{min avg. likelihood ratio per SNP to make a call, this
number is highly dependent on the data type used and should be optimized
by the user based on a sensitivity - specificity trade-off.}

\item{sim_level}{if non-zero, a mix of this level will
be added to the current sample. The contaminant will be randomly
generated from minor allele frequencies of the SNPs in the population.}

\item{baseline}{noise model}

\item{min_depth}{minimum depth for a SNP to be considered by conta}

\item{max_depth}{maximum depth for a SNP to be considered by conta}

\item{loh_cutoff}{minimum maximum likelihood to call a region as LOH}

\item{loh_delta_cutoff}{minimum delta (het deviation) to call a region as LOH}

\item{min_maf}{minimum minor allele frequency to include a SNP}

\item{cf_correction}{cf correction calculated from empirical data}

\item{min_cf}{minimum cf to call}

\item{blackswan}{blackswan term for maximum likelihood estimation}

\item{outlier_frac}{fraction of outlier SNPs (based on depth) to remove}

\item{cores}{number of cores to be used for parallelization}
}
\value{
none
}
\description{
Reads a given counts file, processes it and calculates
contamination likelihood for a set of ranges.
}
