% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{conta_swap}
\alias{conta_swap}
\title{Run conta swap detection}
\usage{
conta_swap(files, labels, concordance_threshold)
}
\arguments{
\item{gt_files}{list of strings, paths to genotype files}

\item{gt_labels}{list of strings, labels for genotype files}

\item{concordance_treshold}{float, threshold for genotype concordance}
}
\value{
data.frame with all pairwise concordances and swap calls
}
\description{
Calculates pairwise concordance between n genotype samples.
Number of files and number of labels must match and be greater
than two. Genotype files must contain gt column. Files and labels
will be paired by index.
This returns a data frame with two metrics for each pair of samples:
concordance and call. Call is true if the concordance is less than
or equal to the concordance threshold. Metrics are only reported
once for each pair of samples.
}
