% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_threshold.R
\name{conta_threshold}
\alias{conta_threshold}
\title{Run the conta thresholding workflow}
\usage{
conta_threshold(sim_data, outdir, extreme_thresh = 1.5,
  quantile_thresh = 0.98, mad_thresh = 3, senss = c(0.95, 0.98, 0.99, 1),
  specs = c(0.95, 0.98, 0.99, 1))
}
\arguments{
\item{sim_data}{Simulation data as read in by `read_sim_data()`.}

\item{outdir}{Path to directory to write outputs to.}

\item{mad_thresh}{Filter out samples that are this many MADs away from the
median, applied after the extreme level and quantile filters.
Defaults to 3 MADs.}

\item{senss}{Vector of sensitivity values.}

\item{specs}{Vector of specificity values.}

\item{extreme_level}{Avg LLR level above which samples will be dropped,
default of 1.5 selected as a very high score
only observed in contaimination positive controls.}

\item{filter_quantile}{Quantile level above which samples will be filtered
to set more aggressive thresholds. Defaults to 0.98,
throwing out the top 2 percent of samples.}
}
\value{
None, writes out TSVs and plots to `outdir`.
}
\description{
Runs the full conta thresholding workflow using a provided set of
simulated contamination data. Workflow consists of filtering outliers
from the data using three filters (see `conta::filter_data()` for details)
followed by calculation of sensitivities, specificities, and LoD at a
variety of cutoffs. Generates tables and diagnostic plots to the specified
output directory.
}
