% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{count_ychr}
\alias{count_ychr}
\title{Get the fraction of bases covered by at least 1 read on Y chr.}
\usage{
count_ychr(bin_file)
}
\arguments{
\item{bin_file}{input tsv file name}
}
\value{
Y chr counts
}
\description{
Input is a bedtools output where the first column shows the chromosome,
and each row stores statistics for a bin of certain size (e.g. 100,000 bp).
7th column in this file represents the percentage of bases on this bin
covered by at least 1 read.
}
