% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{empty_result}
\alias{empty_result}
\title{Return an empty results table to output when quitting early}
\usage{
empty_result(sample_id = "")
}
\arguments{
\item{sample_id}{sample ID to use in empty result.}
}
\description{
Return an empty results table to output when quitting early
}
