% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{exclude_high_loh_regions}
\alias{exclude_high_loh_regions}
\title{Remove regions with high LOH}
\usage{
exclude_high_loh_regions(dat, bin_stats)
}
\arguments{
\item{dat}{data.table SNP data}

\item{bin_stats}{data.table LOH stats for each bin}
}
\value{
data.table data with SNPs on LOH regions removed
}
\description{
Remove regions with high LOH
}
