% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{fail_test}
\alias{fail_test}
\title{Fail if certain conditions are not met}
\usage{
fail_test(dat)
}
\arguments{
\item{dat}{data.table with SNP counts and genotypes}
}
\description{
Fail if certain conditions are not met
}
