% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_threshold.R
\name{filter_data}
\alias{filter_data}
\title{Filter outlier samples from conta simulations}
\usage{
filter_data(sim_data, extreme_level = 1.5, filter_quantile = 0.98,
  mad_thresh = 3)
}
\arguments{
\item{sim_data}{Simulation data as read in by `read_sim_data()`}

\item{extreme_level}{Avg LLR level above which samples will be dropped,
default of 1.5 selected as a very high score
only observed in contaimination positive controls.}

\item{filter_quantile}{Quantile level above which samples will be filtered
to set more aggressive thresholds. Defaults to 0.98,
throwing out the top 2 percent of samples.}

\item{mad_thresh}{Filter out samples that are this many MADs away from the
median, applied after the extreme level and quantile filters.
Defaults to 3 MADs}
}
\value{
Data frame of simulation data with samples flagged as
        outliers removed.
}
\description{
Filters out samples with extreme avg log llr without any simulated
contamination, and the top X percent of samples to set thresholds more
aggresively.
}
