% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_threshold.R
\name{fit_sens_by_conta_level}
\alias{fit_sens_by_conta_level}
\title{Fit a linear model to predict sensitivity from contamination level}
\usage{
fit_sens_by_conta_level(sens_by_level, fname, target_sens = 0.95)
}
\arguments{
\item{sens_by_level}{Data frame of sensitivities at a fixed specificity
split out by contamination level.}

\item{fname}{Path to write the final plot}

\item{target_sens}{Target sensitivity level for LoD estimation}
}
\value{
A list with two elements: model and lod, where `model` is the
        fit object returned by `lm()` and `lod` is the estimated LoD
        for the given target sensitivity.
}
\description{
Fits a polynomial linear regression on log(simulated contamination level)
to predict sensitivity. Writes a plot of the fit as well as returns the
model and an estimated LoD given a target sensitivity.
}
