% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{genotype_concordance}
\alias{genotype_concordance}
\title{Calculate concordance between two samples' genotypes}
\usage{
genotype_concordance(dt1, dt2, min_maf = 0.25)
}
\arguments{
\item{dt1}{genotype set 1}

\item{dt2}{genotype set 2}

\item{min_maf}{minimum population frequency to include a SNP}
}
\description{
Calculate concordance between two samples' genotypes
}
