% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{get_conta_file_loc}
\alias{get_conta_file_loc}
\title{Get conta location from a shiny record}
\usage{
get_conta_file_loc(record, targeted)
}
\arguments{
\item{record}{record needs to have fields Results.Location and SeqSampleID}

\item{targeted}{whether the file set is targeted (vfn) or wgs (gt)}
}
\value{
conta file with genotype and allele frequencies
}
\description{
Get conta location from a shiny record
}
