% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{get_cumulative_sum}
\alias{get_cumulative_sum}
\title{Get cumulative sum of a set of numbers}
\usage{
get_cumulative_sum(v)
}
\arguments{
\item{v}{vector of numbers}
}
\description{
Get cumulative sum of a set of numbers
}
