% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{get_data_tables}
\alias{get_data_tables}
\title{Get data tables for all conta datasets}
\usage{
get_data_tables(mt, i, snps)
}
\arguments{
\item{mt}{swap table to write to}

\item{i}{location on swap table}

\item{snps}{targeted snps data table}
}
\value{
list of conta genotype locations
}
\description{
If randomize is selected, randomly select the entry
}
