% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{get_ethnicity_priors}
\alias{get_ethnicity_priors}
\title{Return CCGA1 ethnicity probabilities}
\usage{
get_ethnicity_priors()
}
\value{
Data frame of CCGA1 ethnicity probabilities
}
\description{
Return CCGA1 ethnicity probabilities
}
