% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{get_exp_cf}
\alias{get_exp_cf}
\title{Return expected contamination fraction}
\usage{
get_exp_cf(cf)
}
\arguments{
\item{cf}{numeric contamination fraction}
}
\value{
expected contamination fraction
}
\description{
Average ratio of heterozygotes for informative (contaminant) SNPs between
100 individual samples was determined to be 80%. For these SNPs, only one
allele (half the fraction) contributes to the contamination. For the
remaining 20% of SNPs which are homozygotes, both alleles contribute to the
contamination.
}
