% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{get_final_mapd}
\alias{get_final_mapd}
\title{Get the final mapd score from CNV qc file}
\usage{
get_final_mapd(cnv_file)
}
\arguments{
\item{cnv_file}{input tsv file name}
}
\value{
Cnv z-score
}
\description{
Get the final mapd score from CNV qc file
}
