% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{get_final_stein}
\alias{get_final_stein}
\title{Get the final stein score from CNV qc file}
\usage{
get_final_stein(cnv_file)
}
\arguments{
\item{cnv_file}{input tsv file name}
}
\value{
Cnv z-score
}
\description{
Get the final stein score from CNV qc file
}
