% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_genomic_seq}
\alias{get_genomic_seq}
\title{Return sequence specified by the chromosome start and end positions}
\usage{
get_genomic_seq(genome_filename = "NA", chr = "NA", start = 0L,
  length = 0L, DEBUG = FALSE)
}
\arguments{
\item{genome_filename}{string name of the genome.fa}

\item{chr}{chromosome name}

\item{start}{sequence start base}

\item{length}{length of sequence}

\item{DEBUG}{print out debug messages}
}
\value{
sequence string
}
\description{
Return sequence specified by the chromosome start and end positions
}
