% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_genomic_seqs}
\alias{get_genomic_seqs}
\title{Return sequences specified by the chromosome start and lengths}
\usage{
get_genomic_seqs(genome_filename, chrs, starts, lengths, DEBUG = FALSE)
}
\arguments{
\item{genome_filename}{string name of the genome.fa}

\item{chrs}{chromosome name}

\item{starts}{sequence start base}

\item{lengths}{length of sequence}

\item{DEBUG}{print out debug messages}
}
\value{
sequence string
}
\description{
Return sequences specified by the chromosome start and lengths
}
