% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{get_genotype_concordance}
\alias{get_genotype_concordance}
\title{Calculate genotype concordance between two samples}
\usage{
get_genotype_concordance(gt1, gt2)
}
\arguments{
\item{gt1}{genotype/variant call table of the host}

\item{gt2}{genotype/variant call table of the contaminant}
}
\value{
numeric fraction of matching genotypes between two samples
}
\description{
Calculate genotype concordance between two samples
}
