% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{get_gt_likelihoods}
\alias{get_gt_likelihoods}
\title{Calculate probabilities per genotype and allele frequency,
based off of Hardy–Weinberg equilibrium.}
\usage{
get_gt_likelihoods(genotype, af)
}
\arguments{
\item{genotype}{genotype, either "0/0", "0/1", or "1/1"}

\item{af}{allele frequency, should be 0 to 1}
}
\value{
genotype likelihood
}
\description{
Calculate probabilities per genotype and allele frequency,
based off of Hardy–Weinberg equilibrium.
}
