% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{get_IUPAC}
\alias{get_IUPAC}
\title{Get ambiguity code for a given combination}
\usage{
get_IUPAC(base1, base2)
}
\arguments{
\item{base1}{first base}

\item{base2}{second base}
}
\value{
iupac code for the combination of bases
}
\description{
TODO: support full IUPAC table:
https://droog.gs.washington.edu/parc/images/iupac.html
}
