% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{get_lr}
\alias{get_lr}
\title{Calculate likelihood ratio for each SNP}
\usage{
get_lr(cf, dat, blackswan = 0.05, outlier_frac = 0.005,
  het_rate = 0.8)
}
\arguments{
\item{cf}{numeric contamination fraction to be tested}

\item{dat}{data.table containing counts and metrics per SNP, hets filtered}

\item{blackswan}{blackswan term for maximum likelihood}

\item{outlier_frac}{fraction of outliers to remove}

\item{het_rate}{ratio of heterozygote to homozygote alt contaminant SNPs}
}
\value{
numeric avg. log-likelihood ratio for the given cf for each SNP
}
\description{
Calculate likelihood ratio for each SNP
}
