% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{get_per_bin_loh}
\alias{get_per_bin_loh}
\title{Calculate per bin (large chromosomal segments) stats}
\usage{
get_per_bin_loh(dat, save_dir, sample, cutoff, blackswan, min_loh = 0.05,
  min_snps = 10)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP}

\item{save_dir}{save directory}

\item{sample}{sample name}

\item{cutoff}{threshold to call loh}

\item{blackswan}{blackswan term sets a limit on probability for each event}

\item{min_loh}{to worry about, if loh is less than this fraction, it is ok}
}
\value{
data.table containing stats per bin
}
\description{
Calculate per bin (large chromosomal segments) stats
}
