% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{get_pregnancy_call}
\alias{get_pregnancy_call}
\title{Make a pregnancy call if there is contamination and host is female}
\usage{
get_pregnancy_call(result, sex, chr_y_stats, chr_y_male_threshold,
  chr_y_female_baseline = 6 * 10^-6, chr_y_male_baseline = 0.001,
  conta_variance = 0.75, min_pregnancy_level = 0.2, mpm = 0.2,
  fpm = 0.6)
}
\arguments{
\item{result}{data frame with conta results}

\item{sex}{estimated sex call for the host}

\item{chr_y_stats}{data frame containing read counts for chrY}

\item{chr_y_male_threshold}{threshold to call male on chrY data}

\item{chr_y_female_baseline}{expected chrY normalized reads for a female}

\item{chr_y_male_baseline}{expected chrY normalized reads for a male}

\item{conta_variance}{expected variance in case of male contamination}

\item{min_pregnancy_level}{minimum contamination fraction for pregnancy}

\item{mpm}{male pregnancy multiplier minimum fraction conta likelihood
expected in the case of a male pregnancy}

\item{fpm}{female_pregnancy_multiplier minimum fraction conta likelihood
expected in the case of a female pregnancy}
}
\value{
pregnancy estimated pregnancy call, i.e. FEMALE, MALE or NA
    (NA is reported in the case of no contamination or host sex not female)
}
\description{
Pregnancy may be defined as a female host with a contamination call that has
less signal on the X chromosome. In general if the child is a male, there
should be no contamination signal on the X chromosome, presence of Y
fragments, and an overall contamination signal. For a female child, there
should be half the expected contamination signal on the X chromosome, and
no presence of Y fragments.
}
