% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{get_records}
\alias{get_records}
\title{Get records for each of the targeted and wgs datasets}
\usage{
get_records(shiny_dt, mt, i, randomize)
}
\arguments{
\item{shiny_dt}{shiny data table to get the records from}

\item{mt}{swap table to write to}

\item{i}{location on swap table}
}
\description{
If randomize is selected, randomly select the entry
}
