% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{get_s3_folders}
\alias{get_s3_folders}
\title{Get folders under a specified s3 directory}
\usage{
get_s3_folders(s3_path)
}
\arguments{
\item{s3_path}{to display the contents}
}
\description{
Returns the set of folders under a given s3 path. get_bucket_df normally
returns all folders and files recursively, this method parses get_bucket_df
output to return only folders directly under the specified dir.
}
