% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{get_sample_name}
\alias{get_sample_name}
\title{Get sample name from gt.loh.tsv file}
\usage{
get_sample_name(gt_loh_file_loc)
}
\arguments{
\item{gt_loh_file_loc}{sample gt file location}
}
\value{
sample id
}
\description{
Get sample name from gt.loh.tsv file
}
