% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{get_sex_call}
\alias{get_sex_call}
\title{Make a sex call for the host}
\usage{
get_sex_call(chr_y_stats, chr_y_male_threshold, result, chr_x_stats = NA,
  chr_x_female_threshold = NA, ftm = 0.75, cf_threshold = 0.2)
}
\arguments{
\item{chr_y_stats}{data frame containing read counts for chrY}

\item{chr_y_male_threshold}{threshold to call male on chrY data}

\item{result}{data frame with conta results}

\item{chr_x_stats}{data frame containing read counts for chrX}

\item{chr_x_female_threshold}{threshold to call female on chrX reads}

\item{ftm}{female threshold multiplier minimum fraction of chrY reads to call
This multiplier defines a gray zone for the sex call.}

\item{cf_threshold}{contamination fraction to skip sex call in case there
is a host switch (used only when there are enough chrY reads). i.e if
both host and contaminant are females, then we can still make a sex call.}
}
\value{
sex estimated sex call, i.e. FEMALE, MALE or NO CALL
}
\description{
Make a sex call for the host
}
\section{TODO}{
 Add a joint caller for chrX and chrY

 Consider adding chrX heterozygous SNP counts to the model

 Report sex chromosome aneuploidies
}

