% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{get_shiny_record}
\alias{get_shiny_record}
\title{Return pre-calculated shiny record}
\usage{
get_shiny_record(shiny_dt, tid, aid, atype)
}
\arguments{
\item{shiny_dt}{shiny data table}

\item{tid}{trimmed patient id with last digit missing}

\item{aid}{CFDNA_CLINICAL or GDNA_CLINICAL}

\item{atype}{wgs or targeted}
}
\value{
matching shiny record
}
\description{
Return pre-calculated shiny record
}
