% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{get_snp_intersection}
\alias{get_snp_intersection}
\title{Create data frame with the intersection of SNPs between sample and vcf}
\usage{
get_snp_intersection(gt_loh_file_loc, prepped_vcf)
}
\arguments{
\item{gt_loh_file_loc}{sample gt.loh.tsv file}

\item{prepped_vcf}{1000G vcf read in and prepped by 'prep_1000g_vcf_af()'}
}
\value{
Data frame with the intersection of gt.loh.tsv file and prepped 1000G vcf
}
\description{
Create data frame with the intersection of SNPs between sample and vcf
}
