% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{get_snp_likelihoods}
\alias{get_snp_likelihoods}
\title{Calculate genotype likelihoods per SNP per 1000G population}
\usage{
get_snp_likelihoods(snp_intersect)
}
\arguments{
\item{snp_intersect}{Data frame of the intersection of snps between sample
and 1000G vcf by 'get_snp_intersection()'}
}
\value{
Data frame with the initital probabilities
}
\description{
Calculate genotype likelihoods per SNP per 1000G population
}
