% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{get_source_lr}
\alias{get_source_lr}
\title{Calculate likelihood ratio based on known source genotypes}
\usage{
get_source_lr(gt1, gt2, cf, blackswan)
}
\arguments{
\item{gt1}{genotype/variant call table of the host}

\item{gt2}{genotype/variant call table of the contaminant}

\item{cf}{numeric contamination fraction}

\item{blackswan}{blackswan term for maximum likelihood}
}
\value{
likelihood ratio from the source
}
\description{
Calculate likelihood ratio based on known source genotypes
}
