% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{init_match_table}
\alias{init_match_table}
\title{Create results table}
\usage{
init_match_table(out_table_name, pairing_table)
}
\arguments{
\item{out_table_name}{location of output file}

\item{pairing_table}{data.table for mapping file}
}
\value{
initialized result table
}
\description{
If it is first time: copy mapping table and add new fields as NA
If it already exists, match it with mapping table to add any new
samples that were added since last time.
}
