% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intersect}
\alias{intersect}
\title{Intersect a TSV file with VCF file and write a TSV output
that contains maf values for each SNP in the TSV file.
In the process, remove multiple-allele SNPs if the option
is turned on.}
\usage{
intersect(tsvFileName, outTsvFileName, vcfFileName, nonDbSnp = FALSE,
  DEBUG = FALSE)
}
\arguments{
\item{tsvFileName}{string name of input TSV file}

\item{outTsvFileName}{string name of output TSV file}

\item{vcfFileName}{string dbsnp VCF file name to be intersected}

\item{DEBUG}{boolean whether to print out messages}
}
\value{
null
}
\description{
Intersect a TSV file with VCF file and write a TSV output
that contains maf values for each SNP in the TSV file.
In the process, remove multiple-allele SNPs if the option
is turned on.
}
