% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intersect_snps}
\alias{intersect_snps}
\title{Intersect a TSV file with VCF file and write a TSV output
that contains maf values for each SNP in the TSV file.
In the process, remove multiple-allele SNPs if the option
is turned on.}
\usage{
intersect_snps(tsv_filename, out_tsv_filename, vcf_filename,
  non_dbSNP = FALSE, DEBUG = FALSE)
}
\arguments{
\item{tsv_filename}{string name of input TSV file}

\item{out_tsv_filename}{string name of output TSV file}

\item{vcf_filename}{string dbsnp VCF file name to be intersected}

\item{DEBUG}{boolean whether to print out messages}
}
\value{
null
}
\description{
Intersect a TSV file with VCF file and write a TSV output
that contains maf values for each SNP in the TSV file.
In the process, remove multiple-allele SNPs if the option
is turned on.
}
