% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{is_compatible_context}
\alias{is_compatible_context}
\title{Test whether the two contexts are compatible with a one base variance in the
middle base}
\usage{
is_compatible_context(context1, context2)
}
\arguments{
\item{context1}{context of the first case}

\item{context2}{context of the second case}
}
\description{
Test whether the two contexts are compatible with a one base variance in the
middle base
}
