% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{l_loh}
\alias{l_loh}
\title{Likelihood of observing the data with LOH}
\usage{
l_loh(ad, dp, maf1, maf2, d = 0, a = 0, er = 3e-04)
}
\arguments{
\item{ad}{alternative allele depth for a given SNP}

\item{dp}{total depth for a given SNP}

\item{maf1}{minor allele frequencies for host sample}

\item{maf2}{minor allele frequencies for candidate contaminant}

\item{d}{LOH level as deviation from 0.5}

\item{a}{alpha contamination level}

\item{er}{error rate}
}
\description{
Calculates the posterior probability of observing a given alternative allele
depth and total depth given a contamination level (alpha) and a loss of
heterozygosity deviation level (delta). Prior probabilities on the host and
contaminate genotypes are calculated using the given minor allele frequencies
for the host (maf1) and contaminate (maf2).
}
