% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{llr_cont}
\alias{llr_cont}
\title{Log likelihood ratio of contamination}
\usage{
llr_cont(ad, depth, maf1, maf2, hm, er, alpha, delta, blackswan)
}
\arguments{
\item{ad}{alternative allele depth for a given SNP}

\item{depth}{total depth for a given SNP}

\item{maf1}{minor allele frequencies for host sample}

\item{maf2}{minor allele frequencies for candidate contaminant}

\item{hm}{het mean frequency from baseline}

\item{er}{error rate for a given SNP}

\item{alpha}{contamination fraction to be tested}

\item{blackswan}{blackswan term sets a limit on probability for each event}

\item{cp}{contamination probability calculated from minor allele frequency}
}
\value{
log likelihood ratio for contamination vs no contamination
}
\description{
At a given level of cf (contamination fraction), calculate
the likelihood of observing the data given ad and depth.
}
