% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{llr_cont_all}
\alias{llr_cont_all}
\title{Log likelihood ratios for contamination for a set of SNPs}
\usage{
llr_cont_all(dat, alpha, blackswan, delta = 0)
}
\arguments{
\item{dat}{data table containing required fields}

\item{alpha}{contamination level level to be tested}

\item{blackswan}{blackswan term sets a limit on probability for each event}
}
\value{
sum of log likelihood ratios for LOH vs no LOH
}
\description{
Average log likelihood ratio across a given set of SNPs
}
