% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{llr_loh_all}
\alias{llr_loh_all}
\title{Log likelihood ratios for loss of heterozygosity for a set of snps}
\usage{
llr_loh_all(dat, delta, blackswan)
}
\arguments{
\item{dat}{data table containing required fields}

\item{delta}{loss of heterozygosity level to be tested}

\item{blackswan}{blackswan term sets a limit on probability for each event}
}
\value{
sum of log likelihood ratios for LOH vs no LOH
}
\description{
Avg log likelihood ratio across a given set of SNPs
}
