% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{load_conta_file}
\alias{load_conta_file}
\title{Load conta files}
\usage{
load_conta_file(conta_loc, snps = NULL)
}
\arguments{
\item{conta_loc}{location of conta vfn or gt file}

\item{snps}{targeted dbsnp data.table}
}
\value{
data.table with conta genotypes
}
\description{
For a given record, mt and snps, read vfn or gt file, and return it.
}
