% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{log_lr}
\alias{log_lr}
\title{Calculate log likelihood ratio}
\usage{
log_lr(cp, depth, cf, mu, ad, blackswan = 0.01, limit_lh = 0.01)
}
\arguments{
\item{cp}{numeric contamination prior based on maf and genotype}

\item{depth}{numeric number of reads}

\item{cf}{numeric tested contamination fraction}

\item{mu}{numeric average error rate for this type of SNP}

\item{ad}{numeric number of alternative alleles called}

\item{blackswan}{numeric limit likelihood of individual hypotheses}

\item{limit_lh}{numeric limit maximum for minimum lh}
}
\value{
numeric log likelihood ratio
}
\description{
Calculate ll-ratio between cf+noise vs noise.
Binomial model for each probability requires average error mu.
}
