% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{logicize_columns}
\alias{logicize_columns}
\title{Logicize columns from shiny data if they are set as string}
\usage{
logicize_columns(shiny_dt, cols)
}
\arguments{
\item{shiny_dt}{data.table}

\item{cols}{columns to logicize}
}
\value{
shiny_dt that is numerized
}
\description{
Logicize columns from shiny data if they are set as string
}
