% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{lr_cont}
\alias{lr_cont}
\title{Log likelihood ratio of contamination}
\usage{
lr_cont(ad, depth, maf, er, alpha, cp, blackswan)
}
\arguments{
\item{ad}{alternative allele depth for a given SNP}

\item{depth}{total depth for a given SNP}

\item{maf}{minor allele frequency for a given SNP}

\item{er}{error rate for a given SNP}

\item{alpha}{contamination fraction to be tested}

\item{blackswan}{blackswan term sets a limit on probability for each event}
}
\value{
log likelihood ratio for contamination vs no contamination
}
\description{
At a given level of cf (contamination fraction), calculate
the likelihood of observing the data given ad and depth.
}
