% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{lr_loh}
\alias{lr_loh}
\title{Log likelihood ratio for loss of heterozygosity hypothesis.}
\usage{
lr_loh(ad, depth, maf, er, delta, blackswan)
}
\arguments{
\item{ad}{alternative allele depth for a given SNP}

\item{depth}{total depth for a given SNP}

\item{maf}{minor allele frequency for a given SNP}

\item{er}{error rate for a given SNP}

\item{delta}{loss of heterozygosity level to be tested}

\item{blackswan}{blackswan term sets a limit on probability for each event}
}
\value{
log likelihood ratio for LOH vs no LOH
}
\description{
At a given level of delta which is the allelic imbalance or LOH level as an
absolute difference from 0.5, calculate the likelihood of observing the
data given a specific ad and depth.
}
