% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_threshold.R
\name{make_roc_plot_by_conta_level}
\alias{make_roc_plot_by_conta_level}
\title{Make ROC curves by conta level}
\usage{
make_roc_plot_by_conta_level(sim_data, annotate_specs = NULL)
}
\arguments{
\item{sim_data}{Simulation data as read in by `read_sim_data()`}

\item{annotate_specs}{Vector of specificity values to draw vertical lines at}
}
\value{
ggplot object containing the ROC plot
}
\description{
Makes a ROC plot with one curve per conta level. Optionally add
vertical dashed lines at given specificty values.
}
