% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{numerize_columns}
\alias{numerize_columns}
\title{Numerize columns from shiny data if they are set as string}
\usage{
numerize_columns(shiny_dt, cols)
}
\arguments{
\item{shiny_dt}{data.table}

\item{cols}{columns to numerize}
}
\description{
Numerize columns from shiny data if they are set as string
}
