% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{optimize_after_grid}
\alias{optimize_after_grid}
\title{optimize further after grid search}
\usage{
optimize_after_grid(dat_chunk, LR_func, blackswan, cf)
}
\arguments{
\item{dat_chunk}{data.table containing counts and metrics per SNP}

\item{LR_func}{Log-likelihood ratio function}

\item{blackswan}{blackswan term sets a limit on probability for each event}
}
\value{
maxima and its objective
}
\description{
optimize further after grid search
}
