% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{optimize_likelihood}
\alias{optimize_likelihood}
\title{Optimize a set of likelihood metrics across SNPs using optimize function.}
\usage{
optimize_likelihood(dat, EE, lr_th, save.dir = NA, sample = NA)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP, hets filtered}

\item{EE}{data.frame error model}

\item{lr_th}{numeric likelihood threshold to make a call}

\item{save.dir}{character location to save the results}

\item{sample}{character sample name}
}
\value{
data.table of likelihoods and related metrics
}
\description{
Optimize a set of likelihood metrics across SNPs using optimize function.
}
