% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{p_cont}
\alias{p_cont}
\title{Probability of observing the data with contamination.}
\usage{
p_cont(ad, depth, maf, er, alpha, cp)
}
\arguments{
\item{ad}{alternative allele depth for a given SNP}

\item{depth}{total depth for a given SNP}

\item{maf}{minor allele frequency for a given SNP}

\item{er}{error rate for a given SNP}

\item{alpha}{contamination fraction}

\item{cp}{contamination probability}
}
\description{
First term signifies the probability of observing the data from a SNP that
has the same genotypes as the host, whereas the second part signifies the
probability of observing the data from an actual contaminated site.
When there is a contamination, it will either manifest as a homozygote allele
coming into an heterozygote site, or as an heterozygote or opposite
homozygote allele coming into the host homozygote site.
}
