% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{p_loh}
\alias{p_loh}
\title{Probability of observing the data with LOH}
\usage{
p_loh(ad, depth, maf, delta = 0, alpha = 0, er = 3e-04)
}
\arguments{
\item{ad}{alternative allele depth for a given SNP}

\item{depth}{total depth for a given SNP}

\item{maf}{minor allele frequency for a given SNP}

\item{delta}{loh level as deviation from 0.5}

\item{alpha}{contamination level}

\item{er}{error rate}
}
\description{
This function tests the probability of observing a given number of alt
reads at a SNP position for a given delta (LOH) and alpha (contamination)
When delta = 0 and alpha = 0, we are essentially testing the probability
of observing the data from a normal site (without LOH or contamination).
}
\details{
The probabilities are conditioned on the genotype. Note the genotype priors
add up to 1. Genotype priors can be one of 3 categories:
1) (1-maf) * (1-maf) is the probability of observing a ref/ref
2) 2 * (1-maf) * maf is the probability of observing a ref/alt, and
3) maf * maf is the probability of observing an alt/alt

Error rate er is a small number, and since the application here
is LOH detection, it should matter only for low coverage (<20) sites where
heterozygosity can be achieved from errors on a homozygote site.
}
