% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{parse_1000g_vcf}
\alias{parse_1000g_vcf}
\title{Read in and parse 1000G vcf}
\usage{
parse_1000g_vcf(vcf, remove_sex_chr = FALSE)
}
\arguments{
\item{vcf}{1000 genomes vcf}

\item{remove_sex_chr}{Defaults to FALSE, meaning sex chromosomes will not be
filtered out. If 'TRUE' sex chromosomes will be removed}
}
\value{
Data frame of parsed vcf
}
\description{
Reads in 1000G vcf file. Parses 'INFO' column into their own columns.
Filters out MNVs (multiple nucleotide variants), Indels, SNPs with no RSID,
and INFO columns of unexpected order.
}
