% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{parse_field}
\alias{parse_field}
\title{Retrieve a specified field from a vcf info string}
\usage{
parse_field(info, field)
}
\arguments{
\item{info}{vcf info string}

\item{field}{a specific field from info}
}
\value{
string field
}
\description{
Retrieve a specified field from a vcf info string
}
