% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{pg}
\alias{pg}
\title{Prior probability of genotype}
\usage{
pg(gt, maf)
}
\arguments{
\item{gt}{0, 1, or 2 corresponding to hom reference, het, hom alternative}

\item{maf}{minor allele frequency}
}
\description{
Prior probability of genotype
}
