% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{pg2}
\alias{pg2}
\title{Prior probability of genotype pair}
\usage{
pg2(gt1, gt2, maf1, maf2)
}
\arguments{
\item{gt1}{Genotype of first sample's SNP
0, 1, or 2 corresponding to hom reference, het, hom alternative}

\item{gt2}{Genotype of second sample's SNP
0, 1, or 2 corresponding to hom reference, het, hom alternative}

\item{maf1}{minor allele frequency first sample's SNP}

\item{maf2}{minor allele frequency of second sample's SNP}
}
\description{
Prior probability of genotype pair
}
