% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_threshold.R
\name{plot_avg_llr_by_conta}
\alias{plot_avg_llr_by_conta}
\title{Make boxplot of conta level vs avg log llr}
\usage{
plot_avg_llr_by_conta(sim_data)
}
\arguments{
\item{sim_data}{Simulation data as read in by `read_sim_data()`}
}
\value{
ggplot object of the plot
}
\description{
Makes a boxplot of conta level vs avg log llr using ggplot
}
