% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_depth_by_chr}
\alias{plot_depth_by_chr}
\title{Plot depth by chromosome}
\usage{
plot_depth_by_chr(dat, save_dir, sample, ext_plot = "depth.png")
}
\arguments{
\item{dat}{data.table with likelihood information per locus}

\item{save_dir}{folder to write out the results}

\item{sample}{sample name to put in filename and plot title}

\item{ext_plot}{extension for the figure output}
}
\value{
none
}
\description{
The goal is to find depth ups and downs
}
