% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_het_dist}
\alias{plot_het_dist}
\title{Plot per bin het ratio ratio across chromosomes}
\usage{
plot_het_dist(dat, save_dir, sample)
}
\arguments{
\item{dat}{data.table with het rates}

\item{save_dir}{character folder to write out the results}

\item{sample}{character sample name to put in filename and plot title}
}
\value{
distortion metric cv of het divergence from 0.5
}
\description{
The goal is to find localized inconcistencies
}
