% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_lr}
\alias{plot_lr}
\title{Calculate a set of likelihood metrics across SNPs}
\usage{
plot_lr(save_dir, sample, dat, per_chr, ext_chr_table = "per_chr.tsv",
  ext_loh_table = "per_bin.tsv", ext_loh_plot = "bin.lr.png",
  min_maf = 0.25)
}
\arguments{
\item{save_dir}{folder location to write files and plots}

\item{sample}{name of the sample}

\item{dat}{data.table with likelihood information per locus}

\item{per_chr}{data.table with per chromosome metrics}

\item{ext_chr_table}{extension for per chr output table}

\item{ext_loh_table}{extension for loh output table}

\item{ext_loh_plot}{extension for loh plot}

\item{min_maf}{minimum allele frequency used to filter SNPs}
}
\value{
none
}
\description{
Calculate a set of likelihood metrics across SNPs
}
