% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_max_likelihood}
\alias{plot_max_likelihood}
\title{Plot log likelihoods and mark the top score}
\usage{
plot_max_likelihood(cf_range, grid_lr, opt_val, save_dir, sample)
}
\arguments{
\item{cf_range}{range of contamination fraction}

\item{grid_lr}{initial grid that was searched}

\item{opt_val}{optimized result}

\item{save_dir}{location to save the results}

\item{sample}{sample name}
}
\value{
none
}
\description{
Each contamination fraction (cf) value that was tested will be plotted
and the top scoring cf highlighted with a red cross. This plot enables
us to see the distribution of log likelihoods.
}
