% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_minor_ratio}
\alias{plot_minor_ratio}
\title{Plot minor allele ratio sorted by chromosomes}
\usage{
plot_minor_ratio(dat, save_dir, sample, ext_plot = "vr.png", maxp = 20000,
  seed = 1)
}
\arguments{
\item{dat}{data.table with snp allele ratio information per locus}

\item{save_dir}{folder to write out the results}

\item{sample}{sample name to put in filename and plot title}

\item{ext_plot}{extension for the figure output}

\item{maxp}{maximum number of points to plot}

\item{seed}{random seed}
}
\value{
none
}
\description{
Plot minor allele ratio sorted by chromosomes
}
