% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{plot_ranked_predictions_per_chromosome}
\alias{plot_ranked_predictions_per_chromosome}
\title{Make tile plot of ranked predictions per chromosome}
\usage{
plot_ranked_predictions_per_chromosome(ranked_predictions)
}
\arguments{
\item{ranked_predictions}{Data frame of ranked predictions by `calculate_ranked_predictions()`}
}
\value{
ggplot object of the plot
}
\description{
Makes tile plot of ranked predictions per chromosome using ggplot
}
