% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_sorted_lr}
\alias{plot_sorted_lr}
\title{Plot sorted log likelihoods}
\usage{
plot_sorted_lr(dat, save_dir, sample)
}
\arguments{
\item{dat}{data.table with likelihood information per locus}

\item{save_dir}{character folder to write out the results}

\item{sample}{character sample name to put in filename and plot title}
}
\value{
none
}
\description{
Plot sorted log likelihoods
}
