% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_ethnicity.R
\name{prep_1000g_vcf_af}
\alias{prep_1000g_vcf_af}
\title{Spike in or remove an allele into a parsed 1000G vcf based on population allele frequency}
\usage{
prep_1000g_vcf_af(parsed_vcf)
}
\arguments{
\item{parsed_vcf}{parsed 1000G vcf generated from 'parse_1000g_vcf()'}
}
\value{
Data frame of parsed vcf
}
\description{
Spikes in an allele count or removes an allele count to SNPs with a 0 or 1
population allele frequency respectively so to not drive log likelihoods
to -Inf or 0.
}
