% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{ratio_and_counts}
\alias{ratio_and_counts}
\title{Add in basic counts and ratios}
\usage{
ratio_and_counts(dat)
}
\arguments{
\item{dat}{data.frame containing counts and metrics per SNP}
}
\value{
data.table containing counts and metrics per SNP
}
\description{
Add in basic counts and ratios
}
