% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{read_data_table}
\alias{read_data_table}
\title{Read a data.table from given path}
\usage{
read_data_table(file, header = TRUE, sep = "\\t", stop_if_missing = FALSE,
  ...)
}
\arguments{
\item{file}{file name to read}

\item{header}{logical whether to read the header}

\item{sep}{the separator}

\item{stop_if_missing}{logical whether to stop execution if file is missing}

\item{...}{additional arguments}
}
\value{
data.table
}
\description{
Check if file name starts with s3 or not, use appropriate function to read.
If the file does not exist, exit with an error message if needed.
}
