% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{read_shiny_data}
\alias{read_shiny_data}
\title{Read shiny data to a table}
\usage{
read_shiny_data(shiny_loc)
}
\arguments{
\item{shiny_loc}{s3 or local path to shiny table}
}
\value{
data.table
}
\description{
Read shiny data to a table
}
