% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{read_sim_results}
\alias{read_sim_results}
\title{Read simulation results from provided path}
\usage{
read_sim_results(file)
}
\arguments{
\item{file}{file name to read}
}
\value{
data.table
}
\description{
Reads conta simulation results from a local or S3 path. Checks
that the required columns are present and properly formatted,
and returns the simulation data as a data frame.
}
