% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{read_vcf_dt}
\alias{read_vcf_dt}
\title{Read a vcf file as a data table}
\usage{
read_vcf_dt(vcf_file, n = 1e+05)
}
\arguments{
\item{vcf_file}{name of the input file}

\item{n}{max number of header lines to skip}
}
\value{
data.table containing vcf data
}
\description{
Read a vcf file as a data table
}
