% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{s3_file_exists}
\alias{s3_file_exists}
\title{Return true if file exists on s3}
\usage{
s3_file_exists(file)
}
\arguments{
\item{file}{character file name to check existence}
}
\description{
Return true if file exists on s3
}
