% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_threshold.R
\name{sens_spec_by_level}
\alias{sens_spec_by_level}
\title{Get senstivity and specificity for each conta level}
\usage{
sens_spec_by_level(sim_data, target_specs = c(0.99))
}
\arguments{
\item{sim_data}{Simulation data as read in by `read_sim_data()`}

\item{target_specs}{Vector of target specificity values}
}
\value{
Data frame with four columns: conta_level, specificity,
        sensitivity, and threshold
}
\description{
For each simulated conta level, get the sensitivty and threshold at
any number of provided specificity values.
}
