% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{set_locations_and_metrics}
\alias{set_locations_and_metrics}
\title{Set result locations and metrics from shiny_dt}
\usage{
set_locations_and_metrics(shiny_dt, mt, i, randomize)
}
\arguments{
\item{shiny_dt}{data.table with shiny results}

\item{mt}{swap table to write to}

\item{i}{location on swap table}

\item{randomize}{whether to randomize the data}
}
\description{
Set result locations and metrics from shiny_dt
}
