% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{set_numeric_chrs}
\alias{set_numeric_chrs}
\title{Set numeric equivalent of chromosomes for sorting purposes}
\usage{
set_numeric_chrs(dat)
}
\arguments{
\item{dat}{data.table to add numeric chromosome}
}
\description{
Set numeric equivalent of chromosomes for sorting purposes
}
\section{TODO}{
 Handle chromosomes outside X, Y and M/MT
}

