% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{set_result_locations}
\alias{set_result_locations}
\title{Get records for each of the targeted and wgs datasets}
\usage{
set_result_locations(mt, i, r)
}
\arguments{
\item{mt}{data.table with swap data}

\item{i}{location on the table}

\item{r}{result_locations for each sample type}
}
\description{
If randomize is selected, randomly select the entry
}
