% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{set_signatures}
\alias{set_signatures}
\title{Set signatures for all datasets}
\usage{
set_signatures(mt, i, d)
}
\arguments{
\item{mt}{swap table to write to}

\item{i}{location on swap table}

\item{d}{list of data tables}
}
\description{
Set signatures for all datasets
}
