% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{sim_conta}
\alias{sim_conta}
\title{Simulate contamination}
\usage{
sim_conta(wgs, cf, seed = 1359)
}
\arguments{
\item{wgs}{data.frame containing counts and metrics per SNP}

\item{cf}{numeric contamination fraction}

\item{seed}{numeric set seed for simulation}
}
\value{
data.frame containing counts and metrics per SNP
}
\description{
It works by adding in each allele proportional to specified
contamination fraction and maf. So it actually simulates a
sample randomly ignoring linkage disequilibrium.
}
