% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_threshold.R
\name{sim_data_to_roc}
\alias{sim_data_to_roc}
\title{Produce a pROC::roc object from simulation data}
\usage{
sim_data_to_roc(sim_data)
}
\arguments{
\item{sim_data}{Dataframe of simulation data as produced by `read_sim_results()`}
}
\value{
A pROC::roc object describing contamination classification from the
        provided simulation results
}
\description{
Returns a pROC::roc object generated from the given simulation data,
expects simulation data in the format returned by `read_sim_results()`.
}
