% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{simulate_loh_conta}
\alias{simulate_loh_conta}
\title{Simulate LOH and/or contamination}
\usage{
simulate_loh_conta(n, min_maf, dp_min, dp_max, er_min, er_max, delta, alpha,
  seed = 1359)
}
\arguments{
\item{n}{number of SNPs to simulate}

\item{min_maf}{minimum maf to simulate (and maximum is 1 - maf)}

\item{dp_min}{min epth to simulate}

\item{dp_max}{max depth to simulate}

\item{er_min}{minimum error rate to simulate}

\item{er_max}{maximum error rate to simulate}

\item{delta}{LOH deviation from heterozygosity}

\item{alpha}{contamination level}

\item{seed}{numeric set seed for simulation}
}
\value{
data.frame containing counts and metrics per SNP
}
\description{
This simulation returns a simple data structure with metrics required to
test LOH functions. It works by simulating 4 chromosomes (2 for host, and
2 for contaminant). It also simulates the ratio of the host chromosomes which
is analogous to loss of heterozygosity.
}
