% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{specify_precision}
\alias{specify_precision}
\title{Returns character list of provided precision}
\usage{
specify_precision(x, precision)
}
\arguments{
\item{x}{numeric data.frame or numeric data.frame column}

\item{precision}{numeric int for decimal place to round to}
}
\description{
This function takes in a given numeric data.frame column, rounds it to the
provided precision and returns back the values as a character list
}
