% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{ssig}
\alias{ssig}
\title{Skewed sigmoid with center around provided mean}
\usage{
ssig(x, m = 0.5)
}
\arguments{
\item{x}{input value}

\item{m}{the value when x = 0}
}
\description{
Skewed sigmoid with center around provided mean
}
