% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{swap}
\alias{swap}
\title{Run swap tool}
\usage{
swap(pairing_table_name, out_table_name, shiny_loc, dbsnp_targeted, corrections,
  randomize = FALSE, cutoff = 0.7, max_conta = 0.2, signatures)
}
\arguments{
\item{pairing_table_name}{name of the mapping table}

\item{out_table_name}{name of the results table to output}

\item{shiny_loc}{location of the shiny database to retrieve metrics}

\item{dbsnp_targeted}{targeted dbsnp vcf file same as used for targeted seq}

\item{corrections}{file with swap corrections that were confirmed}

\item{randomize}{generate random pairs instead of expected ones}

\item{cutoff}{cutoff to make a cfdna targeted vs wgs swap call}

\item{max_conta}{if contamination is above this level, do not call swap}

\item{signatures}{location of file containing signature SNPs}
}
\description{
Compare genotypes across specified sample pairings (specific to GRAIL CCGA
data). Both concordance and distance metrics are reported along with other
informative features such as coverage, CNV indicators and contamination.
}
