% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_threshold.R
\name{thresholds_by_sens_spec}
\alias{thresholds_by_sens_spec}
\title{Get sensitivity and specificity at specified levels}
\usage{
thresholds_by_sens_spec(roc, specs = c(0.9, 0.95, 0.99, 1), senss = c(0.9,
  0.95, 0.99, 1))
}
\arguments{
\item{roc}{Roc object from `pROC::roc`}

\item{specs}{Vector of specificity values}

\item{senss}{Vector of sensitivity values}
}
\value{
Dataframe with 3 columns: specificty, sensitivity, and threshold
}
\description{
Uses a pROC::roc object to calculate specificity, sensitivity,
and the associated threshold at any number of specified specificity
or sensitivity levels. Will find the closest specificity or sensitivity
that is at least as large as the specified value.
}
