% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{update_context}
\alias{update_context}
\title{Update 3-base context for SNPs with alternative homozygote genotypes}
\usage{
update_context(dat)
}
\arguments{
\item{dat}{data.table containing SNPs}
}
\value{
data.table containing SNPs with tagged chunks
}
\description{
Update 3-base context for SNPs with alternative homozygote genotypes
}
