% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{write_data_table}
\alias{write_data_table}
\title{Write a data.table to a given path}
\usage{
write_data_table(out, out_file, header = TRUE, sep = "\\t", ...)
}
\arguments{
\item{out}{data to write}

\item{out_file}{output file to write to}

\item{header}{logical whether to read the header}

\item{sep}{the separator}

\item{...}{additional arguments}
}
\value{
data.table
}
\description{
Check if file name starts with s3 or not, use appropriate function to write.
}
