% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{write_gt_file}
\alias{write_gt_file}
\title{Write genotype file}
\usage{
write_gt_file(dat, max_result, blackswan, outlier_frac, out_file_gt)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP}

\item{max_result}{conta result with maximum likelihood}

\item{blackswan}{blackswan parameter for likelihood calculation}

\item{out_file_gt}{name and location of the output file}

\item{outlier_fract}{outlier fraction}
}
\description{
Genotypes and contamination metrics for each SNP are useful for different
applications such as contamination source and swap detection.
}
