% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{write_results}
\alias{write_results}
\title{Write results to file if any updates were made}
\usage{
write_results(d, mt, out_table_name)
}
\arguments{
\item{d}{list of data tables}

\item{mt}{swap table to write to}

\item{out_table_name}{location to write (either s3 or local)}
}
\description{
Write results to file if any updates were made
}
