% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{add_chunks}
\alias{add_chunks}
\title{Partition each chromosome to bins and tag each SNP with its bin}
\usage{
add_chunks(dat, max_portions = 5)
}
\arguments{
\item{dat}{data.table containing SNPs}

\item{max_portions}{number of portions to split each chromosome}
}
\value{
data.table containing SNPs with tagged chunks
}
\description{
Partition each chromosome to bins and tag each SNP with its bin
}
