% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_error_model.R
\name{add_error_rates}
\alias{add_error_rates}
\title{Add error rates for each base based on ref/ref and alt/alt error rates}
\usage{
add_error_rates(dat, EE, context_mode = FALSE)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP}

\item{EE}{data.frame errror model}

\item{context_mode}{}
}
\value{
data.frame with error model
}
\description{
Add error rates for each base based on ref/ref and alt/alt error rates
}
