% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_parsers.R
\name{annotate_and_filter}
\alias{annotate_and_filter}
\title{Add in more useful columns, annotation and filter sites.}
\usage{
annotate_and_filter(dat, het_limit = 0.25, min_other_ratio = 0.15,
  min_depth = 5, max_depth = 10000, out_frac = 0)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP}

\item{het_limit}{limit fraction to call a heterozygote site.}

\item{min_other_ratio}{If the ratio of highest depth non-SNP allele is at
as high as this number, the SNP will be filtered. These SNPs either have
unusually high error rate (or mutation), or has a hidden (multi-allelic)
SNP. In either case, this position should be filtered from contamination
analysis.}

\item{min_depth}{if a SNP has depth less than this value, it will be
filtered. This metric is mainly for WGS where overall coverage is low.}

\item{max_depth}{if a SNP has depth more than this value, it will be
filtered. This metric is mainly for RNA where some genes have extreme
coverage.}

\item{out_frac}{remove SNPs that are outliers to depth distribution.
This filter is applied post min_depth and max_depth filters to further
remove any outliers that also tend to generate unexpected noise.}
}
\value{
data.frame containing counts and metrics per SNP
}
\description{
Add in more useful columns, annotation and filter sites.
}
