% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{calculate_concordance}
\alias{calculate_concordance}
\title{Calculate concordance for all datasets}
\usage{
calculate_concordance(mt, i, d, ctar_cwgs_cutoff, gtar_gwgs_cutoff,
  ctar_gtar_cutoff, cwgs_gwgs_cutoff)
}
\arguments{
\item{mt}{swap table to write to}

\item{i}{location on swap table}

\item{d}{list of data tables}

\item{ctar_cwgs_cutoff}{cutoff for ctar vs cwgs concordance}

\item{gtar_gwgs_cutoff}{cutoff for gtar vs gwgs concordance}

\item{ctar_gtar_cutoff}{cutoff for ctar vs gtar concordance}

\item{cwgs_gwgs_cutoff}{cutoff for cwgs vs gwgs concordance}
}
\description{
Calculate concordance for all datasets
}
