% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{conta_llr}
\alias{conta_llr}
\title{Conta log likelihood ratio}
\usage{
conta_llr(cf, dat, save_dir = NA, sample = NA, loh = FALSE,
  blackswan = 0.05, outlier_frac = 0.005)
}
\arguments{
\item{cf}{numeric contamination fraction to be tested}

\item{dat}{data.table containing counts and metrics per SNP, hets filtered}

\item{save_dir}{directory to save results}

\item{sample}{name of the sample for printing purposes}

\item{loh}{whether to plot in LOH mode}

\item{blackswan}{blackswan term for maximum likelihood}

\item{outlier_frac}{fraction of outliers to remove}
}
\value{
numeric avg. log-likelihood ratio for the given cf or a more detailed
  result if the save_dir and sample name are specified.
}
\description{
Conta log likelihood ratio
}
