% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_main.R
\name{conta_main}
\alias{conta_main}
\title{Run conta.}
\usage{
conta_main(tsv_file, sample, save_dir, bin_file = NA, cnv_file = NA,
  lr_th = 0.05, sim_level = 0, baseline = NA, cores = 2)
}
\arguments{
\item{tsv_file}{character input tsv file name}

\item{sample}{character name of experiment (basename of file usually)}

\item{save_dir}{character folder where the output will go to}

\item{bin_file}{character input bin file name}

\item{cnv_file}{character input cnv file name}

\item{lr_th}{numeric min avg. likelihood ratio per SNP to make a call, this
number is highly dependent on the data type used and should be optimized
by the user based on a sensitivity - specificity trade-off.}

\item{sim_level}{numeric if non-zero, a mix of this level will
be added to the current sample. The contaminant will be randomly
generated from minor allele frequencies of the SNPs in the population.}

\item{baseline}{data.table noise model}

\item{cores}{number of cores to be used for parallelization}
}
\value{
none
}
\description{
Reads a given counts file, processes it and calculates
contamination likelihood for a set of ranges.
}
