% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_source.R
\name{conta_source}
\alias{conta_source}
\title{Run conta source detection}
\usage{
conta_source(base, out_file, batch_samples = NA, subfolder = "",
  threshold = NA, blackswan = 1, outlier_frac = 0,
  source_threshold = 0.01, cores = 8)
}
\arguments{
\item{base}{character base directory with a batch of conta results}

\item{out_file}{character output file}

\item{batch_samples}{file that contains samples to limit the analysis to}

\item{subfolder}{subfolder name if conta results are in a subfolder}

\item{threshold}{re-call conta based on a new threshold}

\item{blackswan}{blackswan term for maximum likelihood estimation}

\item{outlier_frac}{fraction of outliers to remove}

\item{source_threshold}{difference between MAF llr and genotype llr to call}

\item{cores}{number of cores to use for calculations}
}
\value{
none
}
\description{
Reads a base folder with each conta result under a folder. Each conta run
should contain one *.conta.tsv and *.gt.tsv that correspond to conta
results and genotypes/contaminants of that sample. Conta source detection
will identify the samples for which conta made a call, and then compare its
contaminant reads against genotypes of every other samples' genotypes.
}
\details{
For each genotype comparison, maximum likelihood that is obtained by using
the genotype of the contaminant as the contamination probability
is compared against the original maximum likelihood that is obtained by
using the MAF as the prior. If this likelihood is higher, then the sample
with the highest likelihood is reported as the candidate source of
contamination. Second and third highest candidates are also reported.
}
