% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{genotype_concordance}
\alias{genotype_concordance}
\title{Calculate concordance between two samples' genotypes}
\usage{
genotype_concordance(dt1, dt2)
}
\arguments{
\item{dt1}{data.table of genotype set 1}

\item{dt2}{data.table of genotype set 2}
}
\description{
TODO (edamato): support gt files from other pipelines
}
