% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_loh.R
\name{get_per_bin_loh}
\alias{get_per_bin_loh}
\title{Calculate per bin (large chromosomal segments) stats}
\usage{
get_per_bin_loh(dat, save_dir, sample, min_lr, blackswan, min_loh = 0.15,
  min_snps = 20, max_snps = 200, seed = 1359)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP}

\item{save_dir}{save directory}

\item{sample}{sample name}

\item{min_lr}{likelihood ratio threshold to call LOH}

\item{blackswan}{blackswan term sets a limit on probability for each event}

\item{min_loh}{to worry about affecting contamination results,
if LOH is less than this fraction do not call}

\item{min_snps}{minimum number of snps required to evaluate a region for LOH}

\item{max_snps}{max snps to evaluate in a region, if greater number of
SNPs are present, subsample to this many SNPs.}

\item{seed}{random seed}
}
\value{
data.table containing stats per bin
}
\description{
Calculate per bin (large chromosomal segments) stats
}
