% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_functions.R
\name{get_pregnancy_call}
\alias{get_pregnancy_call}
\title{Make a pregnancy call if there is contamination and host is female}
\usage{
get_pregnancy_call(result, sex, chr_y_stats, chr_y_male_threshold, ftm = 0.1,
  mpm = 0.2, fpm = 0.7)
}
\arguments{
\item{result}{data frame with conta results}

\item{sex}{estimated sex call for the host}

\item{chr_y_stats}{data frame containing read counts for chrY}

\item{chr_y_male_threshold}{threshold to call male on chrY data}

\item{ftm}{female threshold multiplier minimum fraction of chrY reads to call}

\item{mpm}{male pregnancy multiplier minimum fraction conta likelihood
expected in the case of a male pregnancy}

\item{fpm}{female_pregnancy_multiplier minimum fraction conta likelihood
expected in the case of a female pregnancy}
}
\value{
pregnancy estimated pregnancy call, i.e. FEMALE, MALE or NA
    (NA is reported in the case of no contamination or host sex not female)
}
\description{
Pregnancy may be defined as a female host with a contamination call that has
less signal on the X chromosome. In general if the child is a male, there
should be no contamination signal on the X chromosome, presence of Y
fragments, and an overall contamination signal. For a female child, there
should be half the expected contamination signal on the X chromosome, and
no presence of Y fragments.
}
