% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{get_signature}
\alias{get_signature}
\title{Get signature for a given set of SNPs from a conta vfn or gt file}
\usage{
get_signature(conta_dt)
}
\arguments{
\item{conta_dt}{targeted vfn data table}
}
\value{
data.table with conta genotypes
}
\description{
Get signature for a given set of SNPs from a conta vfn or gt file
}
