% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{get_source_llr}
\alias{get_source_llr}
\title{Calculate likelihood ratio based on known source genotypes}
\usage{
get_source_llr(gt1, gt2, cf, blackswan = 1e-06, outlier_frac = 0.005)
}
\arguments{
\item{gt1}{genotype/variant call table of the host}

\item{gt2}{genotype/variant call table of the contaminant}

\item{cf}{numeric contamination fraction}

\item{blackswan}{blackswan term for maximum likelihood}

\item{outlier_frac}{fraction of outliers to remove}
}
\value{
likelihood ratio from the source
}
\description{
Calculate likelihood ratio based on known source genotypes
}
