% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_likelihood.R
\name{optimize_likelihood}
\alias{optimize_likelihood}
\title{Optimize a set of likelihood metrics across SNPs using optimize function.}
\usage{
optimize_likelihood(dat, lr_th, save_dir = NA, filename_prefix = NA,
  loh = FALSE, blackswan, min_cf, cf_correction, outlier_frac)
}
\arguments{
\item{dat}{data.table containing counts and metrics per SNP, hets filtered}

\item{lr_th}{numeric likelihood threshold to make a call}

\item{save_dir}{character location to save the results}

\item{filename_prefix}{character file name prefix}

\item{loh}{whether to plot in LOH mode}

\item{blackswan}{blackswan term for maximum likelihood}

\item{min_cf}{minimum contamination fraction to call}

\item{cf_correction}{cf correction which is subtracted from calculated cf}

\item{outlier_frac}{fraction of outliers to remove}
}
\value{
data.table of likelihoods and related metrics
}
\description{
Optimize a set of likelihood metrics across SNPs using optimize function.
}
