% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_depth_by_chr}
\alias{plot_depth_by_chr}
\title{Plot depth by chromosome}
\usage{
plot_depth_by_chr(dat, save_dir, sample, min_depth, ext_plot = "depth.png",
  max_snps = 10000, seed = 1359)
}
\arguments{
\item{dat}{data.table with likelihood information per locus}

\item{save_dir}{folder to write out the results}

\item{sample}{sample name to put in filename and plot title}

\item{min_depth}{minimum depth to visualize}

\item{ext_plot}{extension for the figure output}

\item{max_snps}{limit to specified number of SNPs for faster plotting}

\item{seed}{random seed}
}
\value{
none
}
\description{
The goal is to find depth ups and downs
}
