% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_lr}
\alias{plot_lr}
\title{Calculate a set of likelihood metrics across SNPs}
\usage{
plot_lr(save_dir, sample, dat, per_chr)
}
\arguments{
\item{save_dir}{folder location to write files and plots}

\item{sample}{name of the sample}

\item{dat}{data.table with likelihood information per locus}

\item{per_chr}{data.table with per chromosome metrics}
}
\value{
none
}
\description{
Calculate a set of likelihood metrics across SNPs
}
