% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_lr_per_bin}
\alias{plot_lr_per_bin}
\title{Plot per bin likelihood across chromosomes}
\usage{
plot_lr_per_bin(dat, save_dir, sample, ext_table = "per_bin.tsv",
  ext_plot = "bin.lr.png")
}
\arguments{
\item{dat}{data.table with likelihood information per locus}

\item{save_dir}{folder to write out the results}

\item{sample}{sample name to put in filename and plot title}

\item{ext_table}{extension for the table output}

\item{ext_plot}{extension for the figure output}
}
\value{
none
}
\description{
The goal is to find localized inconcistencies
}
