% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_lr_per_bin}
\alias{plot_lr_per_bin}
\title{Plot per bin likelihood across chromosomes}
\usage{
plot_lr_per_bin(dat, save_dir, sample)
}
\arguments{
\item{dat}{data.table with likelihood information per locus}

\item{save_dir}{character folder to write out the results}

\item{sample}{character sample name to put in filename and plot title}
}
\value{
none
}
\description{
The goal is to find localized inconcistencies
}
