% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_graphics.R
\name{plot_minor_ratio}
\alias{plot_minor_ratio}
\title{Plot minor allele ratio sorted by chromosomes}
\usage{
plot_minor_ratio(dat, dat_loh = NULL, save_dir, filename_prefix,
  ext_plot = "vr.png", max_snps = 50000, seed = 1)
}
\arguments{
\item{dat}{data.table with snp allele ratio information per locus}

\item{dat_loh}{data.table with snp allele ratio information with LOH removed}

\item{save_dir}{folder to write out the results}

\item{filename_prefix}{file name prefix}

\item{ext_plot}{extension for the figure output}

\item{max_snps}{maximum number of points to plot}

\item{seed}{random seed}
}
\value{
none
}
\description{
Plot minor allele ratio sorted by chromosomes
}
