% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{set_metrics}
\alias{set_metrics}
\title{Set various metrics from shiny to swap table}
\usage{
set_metrics(mt, i, r, min_MaxLh = 0.00025, min_LhDiff = 50)
}
\arguments{
\item{mt}{data.table with swap data}

\item{i}{location on the table}

\item{r}{list of records containing ctar, gtar, cwgs and gwgs results}

\item{min_MaxLh}{minimum MaxLh to make a contamination call}

\item{min_LhDiff}{minimum LhDiff to make a contamination call}
}
\value{
mt with metrics added
}
\description{
Set various metrics from shiny to swap table
}
