% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conta_swap.R
\name{swap}
\alias{swap}
\title{Run swap tool}
\usage{
swap(pairing_table_name, out_table_name, shiny_loc, dbsnp_targeted,
  randomize = FALSE, ctar_cwgs_cutoff = 0.7, gtar_gwgs_cutoff = 0.7,
  ctar_gtar_cutoff = 0.7, cwgs_gwgs_cutoff = 0.7, max_conta = 0.2)
}
\arguments{
\item{pairing_table_name}{name of the mapping table}

\item{out_table_name}{name of the results table to output}

\item{shiny_loc}{location of the shiny database to retrieve sequencing metrics}

\item{dbsnp_targeted}{targeted dbsnp vcf file same as used for targeted seq}

\item{randomize}{generate random pairs instead of expected ones}

\item{ctar_cwgs_cutoff}{cutoff to make a cfdna targeted vs wgs swap call}

\item{gtar_gwgs_cutoff}{cutoff to make a gdna targeted vs wgs swap call}

\item{ctar_gtar_cutoff}{cutoff to make a cfdna targeted vs gdna swap call}

\item{cwgs_gwgs_cutoff}{cutoff to make a wgs targeted vs gdna wgs swap call}

\item{max_conta}{if contamination is above this level, do not call swap}
}
\description{
Compare genotypes across specified sample pairings (specific to GRAIL CCGA
data). Both concordance and distance metrics are reported along with other
informative features such as coverage, CNV indicators and contamination.
}
